code deduplication.
