+++ CommHandler/src/main/java/com/smartgridready/communicator/common/api/values/BooleanValue.java
@@ -3,9 +3,7 @@ package com.smartgridready.communicator.common.api.values;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.BooleanNode;
 
 import java.math.BigInteger;
@@ -98,18 +96,6 @@ public class BooleanValue extends Value {
     @Override
     public JsonNode getJson() {
         return BooleanNode.valueOf(value);
-    }
-
-    @Override
-    public <T> T getJson(Class<T> aClass) {
-        JsonNode node = getJson();
-        ObjectMapper mapper = new ObjectMapper();
-        try {
-            return mapper.treeToValue(node, aClass);
-        } catch (JsonProcessingException e) {
-            var msg = String.format("Unable to map JSON node to the given class '%s'", aClass.getSimpleName());
-            throw new IllegalArgumentException(msg);
-        }
     }
 
     @Override
