+++ CommHandler/src/main/java/com/smartgridready/communicator/common/api/values/Int64UValue.java
@@ -4,9 +4,7 @@ import java.math.BigInteger;
 import java.time.Instant;
 import java.util.Map;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.BigIntegerNode;
 
 public class Int64UValue extends Value {
@@ -104,18 +102,6 @@ public class Int64UValue extends Value {
     @Override
     public JsonNode getJson() {
         return BigIntegerNode.valueOf(value);
-    }
-
-    @Override
-    public <T> T getJson(Class<T> aClass) {
-        JsonNode node = getJson();
-        ObjectMapper mapper = new ObjectMapper();
-        try {
-            return mapper.treeToValue(node, aClass);
-        } catch (JsonProcessingException e) {
-            var msg = String.format("Unable to map JSON node to the given class '%s'", aClass.getSimpleName());
-            throw new IllegalArgumentException(msg);
-        }
     }
 
     @Override
