+++ CommHandler/src/main/java/com/smartgridready/communicator/common/api/values/JsonValue.java
@@ -3,8 +3,6 @@ package com.smartgridready.communicator.common.api.values;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
@@ -158,17 +156,6 @@ public class JsonValue extends Value {
     @Override
     public JsonNode getJson() {
         return value;
-    }
-
-    @Override
-    public <T> T getJson(Class<T> aClass) {
-        ObjectMapper mapper = new ObjectMapper();
-        try {
-            return mapper.treeToValue(value, aClass);
-        } catch (JsonProcessingException e) {
-            var msg = String.format("Unable to map JSON node to the given class '%s'", aClass.getSimpleName());
-            throw new IllegalArgumentException(msg);
-        }
     }
 
     @Override
