+++ CommHandler/src/main/java/com/smartgridready/communicator/common/api/values/Value.java
@@ -4,7 +4,9 @@ import com.smartgridready.ns.v0.DataTypeProduct;
 import com.smartgridready.ns.v0.EnumMapProduct;
 import com.smartgridready.ns.v0.ModbusBoolean;
 import com.smartgridready.ns.v0.ModbusDataType;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.smartgridready.communicator.common.helper.JsonHelper;
 import com.smartgridready.communicator.modbus.helper.ConversionHelper;
 import com.smartgridready.driver.api.common.GenDriverException;
@@ -31,11 +33,21 @@ public abstract class Value  {
     public abstract Map<String, Boolean> getBitmap();
     public abstract Instant getDateTime();
     public abstract JsonNode getJson();
-    public abstract <T> T getJson(Class<T> aClass);
     public abstract void absValue();
     public abstract void roundToInt();
 
     public abstract Value[] asArray();
+
+    public <T> T getJson(Class<T> aClass) {
+        JsonNode node = getJson();
+        ObjectMapper mapper = new ObjectMapper();
+        try {
+            return mapper.treeToValue(node, aClass);
+        } catch (JsonProcessingException e) {
+            var msg = String.format("Unable to map JSON node to the given class '%s'", aClass.getSimpleName());
+            throw new IllegalArgumentException(msg);
+        }
+    }
 
     public int[] toModbusRegister(ModbusDataType modbusDataType) {
